Switch to generated TypedDocument queries
