+++ app/data/mutations.js
@@ -1,0 +1,27 @@
+import { graphql } from 'app/data/gql/gql'
+
+export const newDeckMutation = graphql(/* GraphQL */ `
+  mutation Mutation($objects: [UserDeckInsertInput!]!) {
+    insertIntoUserDeckCollection(objects: $objects) {
+      records {
+        id
+        lang
+      }
+    }
+  }
+`)
+
+export const newCardMutation = graphql(/* GraphQL */ `
+  mutation InsertIntoDeckMembershipCollection(
+    $objects: [DeckMembershipInsertInput!]!
+  ) {
+    insertIntoDeckMembershipCollection(objects: $objects) {
+      records {
+        id
+        status
+        cardPhraseId
+        deckId
+      }
+    }
+  }
+`)
